refactor: reorganize SoundData type imports and rename buttonPreset to presetButtons for consistency
