+++ main.ts
@@ -4,8 +4,9 @@ import { app, BrowserWindow, type IpcMainEvent, ipcMain } from 'electron'
 import { defaultSettings } from '@/constants/settings'
 import { setIsQuitting } from '@/store/quitting'
 import Store from '@/store/settings'
-import type { HotkeyMap, SoundData } from '@/types'
+import type { HotkeyMap } from '@/types'
 import type { Settings } from '@/types/settings'
+import type { SoundData } from '@/types/sound'
 import { convertOpus } from '@/utils/audio/convertOpus'
 import { soundsManager } from '@/utils/sound/manager'
 import { createWindow, win } from '@/window/main'
