+++ src/components/sounds/button.tsx
@@ -5,7 +5,7 @@ import { Button } from '@/components/ui/button'
 import { Checkbox } from '@/components/ui/checkbox'
 import { Popover, PopoverContent, PopoverTrigger } from '@/components/ui/popover'
 import { Separator } from '@/components/ui/separator'
-import { buttonPreset } from '@/constants/themes'
+import { presetButtons } from '@/constants/themes'
 import { useAudio } from '@/context/audio'
 import { useSettings } from '@/context/setting'
 import { useSounds } from '@/context/sounds'
@@ -101,7 +101,7 @@ export function SoundButton({
                   </>
                 ))}
               <div className="grid grid-cols-5 gap-2">
-                {buttonPreset.map((presetColor) => (
+                {presetButtons.map((presetColor) => (
                   <button
                     type="button"
                     key={presetColor}
