+++ src/hooks/useHotkey.ts
@@ -1,6 +1,7 @@
 import { useCallback, useEffect, useState } from 'react'
 
-import type { HotkeyMap, SoundData } from '@/types'
+import type { HotkeyMap } from '@/types'
+import type { SoundData } from '@/types/sound'
 
 export function useHotkey(_soundData: SoundData[], onSoundPlay: (soundId: string) => void) {
   const [hotkeyMap, setHotkeyMap] = useState<HotkeyMap>({})
