+++ src/types/global.d.ts
@@ -1,0 +1,7 @@
+import { IpcApi } from '@/types'
+
+declare global {
+  interface Window {
+    electronAPI: IpcApi
+  }
+}
