+++ src/types/sound.ts
@@ -1,0 +1,8 @@
+export interface SoundData {
+  id: string
+  file: string
+  title: string
+  frequent?: boolean
+  isUserAdded?: boolean
+  format?: 'opus' | 'mp3'
+}
