+++ src/utils/audio/addTrack.ts
@@ -2,14 +2,14 @@ import { createId } from '@/utils/audio/createId'
 
 export async function addTrack(file: File, type: 'sound' | 'music', customTitle?: string) {
   try {
-    const soundData = await createId(file, type, customTitle)
+    const sound = await createId(file, type, customTitle)
 
     await window.electronAPI.addSound({
-      sound: soundData,
+      sound,
       type,
     })
 
-    return soundData
+    return sound
   } catch (error) {
     console.error('Failed to add sound:', error)
     throw error
