+++ src/utils/audio/createId.ts
@@ -1,4 +1,4 @@
-import type { SoundData } from '@/types'
+import type { SoundData } from '@/types/sound'
 import { generateId } from '@/utils/sound/generateId'
 
 export const createId = async (file: File, type: 'sound' | 'music', customTitle?: string) => {
