+++ src/utils/sound/getSound.ts
@@ -1,5 +1,5 @@
 import { rawSoundData } from '@/constants/audio'
-import type { SoundData } from '@/types'
+import type { SoundData } from '@/types/sound'
 import { generateId } from '@/utils/sound/generateId'
 
 export const getSound = (): SoundData[] => {
